backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
motor==3.7.0
pydantic-settings==2.8.1
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
python-multipart==0.0.20
pymupdf==1.25.5
openai==1.75.0
